DONE:- add filter in toggle api

+++ src/controller/profile_controller.js
@@ -1,33 +1,40 @@
-const ApiError=require("../utils/error")
-const { EMPLOYEE_ROLE } = require("../config/string")
-const EmployeeModel=require("../model/employee_model")
+const ApiError = require("../utils/error")
+const EmployeeModel = require("../model/employee_model")
 const cloudinary = require("../utils/cloudinary");
 const fs = require("fs");
-const { FOLDER_NAME } = require("../config/string");
 
-async function addProfile(req,res,next){
+async function editProfile(req, res, next) {
     try {
-        console.log(req.file);
         const path = req.file.path;
-        const result = await cloudinary.uploader.upload(path);
-        req.body.profilePic = result.secure_url;
-        req.body.publicId = result.public_id;
-        fs.unlinkSync(path);
-        const id=req.id
-        const profile=await EmployeeModel.findByIdAndUpdate({_id:id},{$set:req.body},{new:true})
-        res.status(200).json({success:true,data:profile,message:"profile details added successfully"})
+        if(path) {
+            const result = await cloudinary.uploader.upload(path);
+            req.body.profilePic = result.secure_url;
+            req.body.publicId = result.public_id;
+            fs.unlinkSync(path);
+        }
+        const id = req.id;
+        const profile = await EmployeeModel.findByIdAndUpdate({ _id: id }, { $set: req.body }, { new: true });
+        res.status(200).json({ success: true, data: profile, message: "profile details update successfully" });
     } catch (e) {
-        next(new ApiError(400, e.message))
+        next(new ApiError(400, e.message));
     }
 }
 
-async function getProfile(req,res,next){
+async function addOrUpdateProfile(req, res, next) {
     try {
-        const id=req.id
-        const profile=await EmployeeModel.findById({_id:id})
-        res.status(200).json({success:true,data:profile})
+        const path = req.file.path;
+        if (path) {
+            const result = await cloudinary.uploader.upload(path);
+            req.body.profilePic = result.secure_url;
+            req.body.publicId = result.public_id;
+            fs.unlinkSync(path);
+        }
+        const id = req.params.id;
+        const profile = await EmployeeModel.findByIdAndUpdate({ _id: id }, { $set: req.body }, { new: true });
+        res.status(200).json({ statusCode: 200, success: true, data: profile, message: "employee Updated successfully" });
     } catch (e) {
-         next(new ApiError(400,e.message))
+        next(new ApiError(400, e.message));
     }
 }
-module.exports={addProfile}
+
+module.exports = { editProfile, addOrUpdateProfile };
